requirements.txt
requests
python-dotenv
pillow
geocoder